LayoutTests/fast/mediastream/MediaStream-video-element-expected.txt

Tests MediaStreamTrack callbacks.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".




**** calling mediaDevices.getUserMedia() ****
PASS mediaDevices.getUserMedia succeeded.

**** setup video element ****
video.src = window.URL.createObjectURL(mediaStream)
Event 'loadedmetadata'
Event 'loadeddata'
Event 'canplay'

*** start playback ****
video.play()
Event 'play'
Event 'canplaythrough'
video.pause()

**** check video element ****
PASS video.videoWidth is 640
PASS video.videoHeight is 480
PASS video.src.indexOf("blob:") is 0
PASS video.srcObject is mediaStream

**** check video tracks ****
PASS video.videoTracks.length is 1
PASS video.videoTracks[0].id is mediaStream.getVideoTracks()[0].id
video.videoTracks[0] properties:
  track.kind = main
  track.language = 
  track.id = <UUID>
  track.label = Mock video device
  track.selected = true
  track.sourceBuffer = null

**** check audio tracks ****
PASS video.audioTracks.length is 1
PASS video.audioTracks[0].id is mediaStream.getAudioTracks()[0].id
video.audioTracks[0] properties:
  track.kind = main
  track.language = 
  track.id = <UUID>
  track.label = Mock audio device
  track.enabled = true
  track.sourceBuffer = null

PASS successfullyParsed is true

TEST COMPLETE
